Move tests to `scalacheck.tests` package.
