+++ core/jvm/src/test/scala/scalacheck/tests/ShrinkSpecificationJVM.scala
@@ -1,0 +1,40 @@
+/*
+ * ScalaCheck
+ * Copyright (c) 2007-2021 Rickard Nilsson. All rights reserved.
+ * http://www.scalacheck.org
+ *
+ * This software is released under the terms of the Revised BSD License.
+ * There is NO WARRANTY. See the file LICENSE for the full text.
+ */
+
+package scalacheck.tests
+
+import org.scalacheck.Properties
+import org.scalacheck.Prop.{forAll, propBoolean}
+import org.scalacheck.Shrink.shrink
+import ShrinkSpecification.shrinkClosure
+
+/** @todo
+  *   should work not only JVM but also Scala.js
+  */
+object ShrinkSpecificationJVM extends Properties("Shrink JVM") {
+
+  property("list") = forAll { (l: List[Int]) =>
+    !shrink(l).contains(l)
+  }
+
+  property("non-empty list") = forAll { (l: List[Int]) =>
+    (!l.isEmpty && l != List(0)) ==> {
+      val ls = shrinkClosure(l)
+      ls.toList.toString |: (ls.contains(Nil) && ls.contains(List(0)))
+    }
+  }
+
+  property("xmap vector from list") = forAll { (v: Vector[Int]) =>
+    (!v.isEmpty && v != Vector(0)) ==> {
+      val vs = shrinkClosure(v)
+      Vector(vs: _*).toString |: (vs.contains(Vector.empty) && vs.contains(Vector(0)))
+    }
+  }
+
+}
