+++ core/jvm/src/test/scala/scalacheck/tests/time/ShrinkSpecification.scala
@@ -1,0 +1,26 @@
+/*
+ * ScalaCheck
+ * Copyright (c) 2007-2021 Rickard Nilsson. All rights reserved.
+ * http://www.scalacheck.org
+ *
+ * This software is released under the terms of the Revised BSD License.
+ * There is NO WARRANTY. See the file LICENSE for the full text.
+ */
+
+package scalacheck.tests.time
+
+import org.scalacheck.Prop._
+import org.scalacheck.Shrink._
+import org.scalacheck._
+
+import java.time._
+
+object ShrinkSpecification extends Properties("time.Shrink") {
+  property("shrink[Duration]") = forAll { (n: Duration) =>
+    !shrink(n).contains(n)
+  }
+
+  property("shrink[Period]") = forAll { (n: Period) =>
+    !shrink(n).contains(n)
+  }
+}
