+++ core/shared/src/test/scala-2.13+/scalacheck/tests/ScalaVersionSpecific.scala
@@ -1,0 +1,14 @@
+/*
+ * ScalaCheck
+ * Copyright (c) 2007-2021 Rickard Nilsson. All rights reserved.
+ * http://www.scalacheck.org
+ *
+ * This software is released under the terms of the Revised BSD License.
+ * There is NO WARRANTY. See the file LICENSE for the full text.
+ */
+
+package scalacheck.tests
+
+private[tests] object ScalaVersionSpecific
+
+private[tests] trait GenSpecificationVersionSpecific
